VoxAI_Chat_API/requirements.txt
llama-cpp-python
psutil
